requirements.txt
pyspark==3.4.4
delta-spark==2.4.0
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
